Chore: Split the UI code inside main.go into a separate package ui
